requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
PyJWT==2.8.0
pytest==7.4.2
